feat: move utils to utils.go
